refactor(Application): 차를 생성하는 책임을 Application에서 CarFactory로 분리한다.
