+++ src/main/java/racingcar/Application.java
@@ -1,11 +1,16 @@
 package racingcar;
 
 import camp.nextstep.edu.missionutils.Console;
+import racingcar.domain.Car;
+import racingcar.domain.CarFactory;
+
+import java.util.List;
 
 public class Application {
     public static void main(String[] args) {
         // TODO 구현 진행
-        String[] carNames = getCarNames();
+
+        List<Car> cars = CarFactory.manufactureByUserInput();
         int trialInput = getTrial();
     }
 
@@ -19,15 +24,7 @@ public class Application {
         return Integer.parseInt(trialInput);
     }
 
-    private static String[] getCarNames() {
-        String[] carNames;
-        do {
-            System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
-            String carNamesInput = Console.readLine();
-            carNames = carNamesInput.split(",");
-        } while (isIllegalName(carNames));
-        return carNames;
-    }
+
 
     private static boolean isIllegalTrial(String trialInput) {
         try {
@@ -56,21 +53,4 @@ public class Application {
 
     }
 
-    private static boolean isIllegalName(String[] seperatedNames) {
-        for (String name : seperatedNames) {
-            try {
-                assertName(name);
-            } catch (IllegalArgumentException e) {
-                System.out.println(e.getMessage());
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private static void assertName(String name) {
-        if (name.length() > 5) {
-            throw new IllegalArgumentException("[ERROR] 자동차 이름의 크기는 5를 초과할 수 없습니다. 현재 자동차 이름 = " + name);
-        }
-    }
 }
